requirements.txt
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.3
djangorestframework==3.15.2
drf-yasg==1.21.8
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
inflection==0.5.1
mysqlclient==2.2.6
packaging==24.2
python-dotenv==1.0.1
python-telegram-bot==21.7
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
loguru==0.7.2
